refactor: complete legacy code archival — remove storage_backend branching
- Remove speech_storage_backend parameter and Union type from CorpusLoader;
  _load_repository() now unconditionally builds SpeechStore + SpeechRepositoryFast
- Drop api_swedeb.legacy.speech_lookup import from corpus_loader.py
- Remove speech.storage_backend key from config/config.yml and tests/config.yml
- Delete test_speech_parity.py (superseded by full corpus parity check)
- Delete test_speech_repository_fast.py; replace with test_speech_repository.py
  (all legacy_repo fixtures and dual-backend parity/benchmark tests removed;
   fast-backend behavior, reliability, and benchmark coverage retained)
- Update test_corpus_loader.py: remove SpeechTextRepository import and legacy
  backend test; mock SpeechStore/SpeechRepositoryFast; add skipif guard for
  integration test that requires a built bootstrap_corpus on disk
- Update .github/copilot-instructions.md and AGENTS.md to remove legacy|prebuilt
  rollout language; CorpusLoader now exclusively uses the prebuilt backend
- Update tests/README.md: remove pre-existing-failures note
- Mark all cleanup checklist items as complete in MERGED_SPEECH_CORPUS_IMPLEMENTATION_PLAN.md

Closes legacy archival checklist. 515 unit tests pass.
