+++ config/config.yml
@@ -75,7 +75,6 @@ vrt:
   tag: text
 
 speech:
-  storage_backend: legacy  # legacy | prebuilt
   bootstrap_corpus_folder: /data/swedeb/v1.4.1/speeches/bootstrap_corpus
 
 kwic:
